Update BsonDocumentTest
I've found a potential bug in the default BsonDocument implementation
DefaultDocument. The default equals method (which is delegated to a
LinkedHashMap if my memory serves me right) compares arrays by reference.

I've added a test case for this which fails like it should. When I'll have time
I'll write a proper hashCode/equals for DefaultDocument.